data-getter/requirements.txt
astroid==2.1.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.6.1
PyGithub==1.40
PyJWT==1.6.4
pylint==2.2.2
requests==2.21.0
six==1.12.0
typed-ast==1.2.0
urllib3==1.23
wrapt==1.11.0
